Complete collecting trace for local.get
