+++ src/trace/mod.rs
@@ -1,3 +1,5 @@
 pub mod state_trace_tuple;
 pub mod storage_access_record;
 pub mod state_trace_manager;
+pub mod section_type;
+pub mod storage_type;
