+++ src/trace/storage_type.rs
@@ -1,0 +1,6 @@
+#[derive(Debug)]
+pub enum StorageType {
+    Memory = 0,
+    Stack = 1,
+    Undefined = 2,
+}
